FractionalDiffEq.jl APIs

Problem definition:

FractionalDiffEq.FDEProblem — Type

FDEProblem(f, α, u0, T, h)

Define a Fractional Differential in time interval [0, T] with initial value y(0)=u₀, α-order derivative and step size h.

General solve API

FractionalDiffEq.solve — Function

solve(FODEProblem, PECE())

After define the FDEProblem, use PECE(Predict-Evaluate-Correct-Evaluate) algorithm to computing the Fractional Differential Equation

solve(equation, right, h, T, MatrixDiscrete())

Using the Matrix Discretization algorithm proposed by Prof Igor Podlubny to obtain the numerical solution.

solve(α, β, T, M, N, FPDEMatrixDiscrete())

When using the Martix

Current algorithms:

Base type

FractionalDiffEq.FractionalDiffEqAlgorithm — Type

Base type of FractionalDiffEq algorithms

Problem types

FractionalDiffEq.FODEProblem — Type

FODEProblem(f, α, u0, T, h)

Fractional Ordinary Differential Equations definition

FractionalDiffEq.FPDEProblem — Type

FPDEProblem()

Detailed types

FractionalDiffEq.PECE — Type

Predict-Evaluate-Correct-Evaluate algorithm.

For more details, please refer to Predictor-Corrector algorithms

This PECE algorithm is taken from Diethelm's paper.

@article{
title={A predictor-corrector approach for the numerical solution of fractional differential equations},
author={Diethelm, Kai and Ford, Neville J. and Freed, Alan D.}
doi={https://doi.org/10.1023/A:1016592219341}
}

FractionalDiffEq.FODEMatrixDiscrete — Type

Using triangular strip matrices to discrete fractional ordinary differential equations to simple algebra system and solve the system.

FractionalDiffEq.FPDEMatrixDiscrete — Type

Using triangular strip matrices to discrete fractional partial differential equations to simple algebra system and solve the system.

Some models

FractionalDiffEq.bagleytorvik — Function

bagleytorvik(p1, p2, p3, T, h)

By specifying the parameters of Bagley Torvik Equation, we can use bagleytorvik to directly obtain the numerical approximation of a bagley torvik equation.

p2 ≠ 0

Please note that the parameter of fractional derivative part must not be 0

FractionalDiffEq.diffusion — Function

diffusion(α, β)

Diffusion equation with time fractional derivative.

α and β are the coefficients of the time derivative and spatial derivative.