FractionalDiffEq.jl APIs

Problem definition:

FractionalDiffEq.FDEProblem — Type

FDEProblem

General parent type for all kinds of problems in FractionalDiffEq.jl.

General solve API

FractionalDiffEq.solve — Function

solve(problem, args..., Alg())

Generalproblem solving API for solving FDE problems.

The general solving API solve can accept various kinds of inputs, including FDEProblem, FODEProblem and FPDEProblem

Current algorithms:

Base type

FractionalDiffEq.FractionalDiffEqAlgorithm — Type

Base type of FractionalDiffEq algorithms

Problem types