FractionalDiffEq.jl APIs

Problem definition:

FractionalDiffEq.FDEProblem — Type

FDEProblem

General parent type for all kinds of problems in FractionalDiffEq.jl.

General solve API

FractionalDiffEq.solve — Function

solve(FODEProblem, PECE())

After define the FDEProblem, use PECE(Predict-Evaluate-Correct-Evaluate) algorithm to computing the Fractional Differential Equation

solve(equation, right, h, T, MatrixDiscrete())

Using the Matrix Discretization algorithm proposed by Prof Igor Podlubny to obtain the numerical solution.

References

@inproceedings{Podlubny1998FractionalDE, title={Fractional differential equations}, author={Igor Podlubny}, year={1998} }

solve(α, β, T, M, N, FPDEMatrixDiscrete())

When using the Martix

solve(parameters, order, lparameters, lorders, u, t)

Use Closed-Form Hankel matrix solution to obtain numerical solution at zero initial condition.

solve(parameters, order, lparameters, lorders, u, t)

Use Closed-Form solution to obtain numerical solution at zero initial condition.

The general solving API solve can accept various kinds of inputs, including FDEProblem, FODEProblem and FPDEProblem

Current algorithms:

Base type

FractionalDiffEq.FractionalDiffEqAlgorithm — Type

Base type of FractionalDiffEq algorithms

Problem types

FractionalDiffEq.SingleTermFODEProblem — Type

SingleTermFODEProblem(f, α, h)

Define a single term fractional ordinary differential equation, there are only one term in this problem.

FractionalDiffEq.MultiTermsFODEProblem — Type

MultiTermsFODEProblem(parameters, orders, rparameters, rorders)

Define a multi-terms fractional ordinary differential equation.

FractionalDiffEq.FPDEProblem — Type

FPDEProblem(α, β, T, M, N)

Detailed types

FractionalDiffEq.PECE — Type

Predict-Evaluate-Correct-Evaluate algorithm.

For more details, please refer to Predictor-Corrector algorithms

This PECE algorithm is taken from Diethelm's paper.

@article{
title={A predictor-corrector approach for the numerical solution of fractional differential equations},
author={Diethelm, Kai and Ford, Neville J. and Freed, Alan D.}
doi={https://doi.org/10.1023/A:1016592219341}
}

FractionalDiffEq.FODEMatrixDiscrete — Type

Using triangular strip matrices to discrete fractional ordinary differential equations to simple algebra system and solve the system.

@inproceedings{Podlubny2000MATRIXAT, title={MATRIX APPROACH TO DISCRETE FRACTIONAL CALCULUS}, author={Igor Podlubny}, year={2000} }

FractionalDiffEq.FPDEMatrixDiscrete — Type

Using triangular strip matrices to discrete fractional partial differential equations to simple algebra system and solve the system.

@article{2009, title={Matrix approach to discrete fractional calculus II: Partial fractional differential equations}, DOI={10.1016/j.jcp.2009.01.014}, author={Podlubny, Igor and Chechkin, Aleksei and Skovranek, Tomas and Chen, YangQuan and Vinagre Jara, Blas M.}, }

Some models

FractionalDiffEq.bagleytorvik — Function

bagleytorvik(p1, p2, p3, T, h)

By specifying the parameters of Bagley Torvik Equation, we can use bagleytorvik to directly obtain the numerical approximation of a bagley torvik equation.

p2 ≠ 0

Please note that the parameter of fractional derivative part must not be 0

FractionalDiffEq.diffusion — Function

diffusion(α, β)

Diffusion equation with time fractional derivative.

α and β are the coefficients of the time derivative and spatial derivative.