FractionalDiffEq.jl APIs

Problem definition:

FractionalDiffEq.FDEProblem — Type

FDEProblem(f, α, u0, T, h)

Define a Fractional Differential in time interval [0, T] with initial value y(0)=u₀, α-order derivative and step size h.

General solve API

FractionalDiffEq.solve — Function

solve(FDEProblem, PECE())

After define the FDEProblem, use PECE(Predict-Evaluate-Correct-Evaluate) algorithm to computing the Fractional Differential Equation

Current algorithms:

Base type

FractionalDiffEq.FractionalDiffEqAlgorithm — Type

Base type of FractionalDiffEq algorithms

Detailed types

FractionalDiffEq.PECE — Type

Predict-Evaluate-Correct-Evaluate algorithm.

For more details, please refer to Predictor-Corrector algorithms

This PECE algorithm is taken from Diethelm's paper.

@article{
title={A predictor-corrector approach for the numerical solution of fractional differential equations},
author={Diethelm, Kai and Ford, Neville J. and Freed, Alan D.}
doi={https://doi.org/10.1023/A:1016592219341}
}

FractionalDiffEq.MatrixDiscrete — Type

Using triangular strip matrices to discrete fractional differential equations to simple algebra system and solve the system.

@inproceedings{Podlubny2000MATRIXAT,
  title={MATRIX APPROACH TO DISCRETE FRACTIONAL CALCULUS},
  author={Igor Podlubny},
  year={2000}
}