Problem Types

Problem Types

FODE Problem

FPDE Problem

FDDE Problem

DODE Problem

Fractional Difference Problem

FIE Problem

The general fractional differential equations problem type:

FractionalDiffEq.FDEProblem — Type

FDEProblem

General type for all kinds of problems in FractionalDiffEq.jl.

FODE Problem

Fractional ordinary problems type, we can classified them as Single-Term and Multi-Term problems:

\[D^{\alpha}y(t)=f(t, y)\]

\[\sum_{s=0}^pc_sD^{\beta_s}y=f\]

FractionalDiffEq.SingleTermFODEProblem — Type

SingleTermFODEProblem(f, α, u0, tspan)

Define a single term fractional ordinary differential equation, there are only one fractional differential operator in this problem.

FractionalDiffEq.MultiTermsFODEProblem — Type

MultiTermsFODEProblem(parameters, orders, rightfun, u0, T)

MultiTermsFODEProblem(parameters, orders, rightfun, rparameters, rorders, u0, T)

Define a multi-terms fractional ordinary differential equation.

FractionalDiffEq.FODESystem — Type

FODESystem(f, α, u0, tspan, p)

Define system of fractional differential equations.

FFODE Problem

FractionalDiffEq.FFPODEProblem — Type

FFPODEProblem(f, α, u0, tspan, p)

Define fractal-fractional differential equations problems with power law kernel.

FractionalDiffEq.FFEODEProblem — Type

FFEODEProblem(f, α, u0, tspan, p)

Define fractal-fractional differential equations problems with expoenntial decay kernel.

FractionalDiffEq.FFMODEProblem — Type

FFMODEProblem(f, α, u0, tspan, p)

Define fractal-fractional differential equations problems with Mittag Leffler kernel.

FPDE Problem

Missing docstring.

Missing docstring for FractionalDiffEq.FPDEProblem. Check Documenter's build log for details.

FDDE Problem

FractionalDiffEq.FDDEProblem — Type

FDDEProblem(f, ϕ, α, τ, tspan)

f: The function describing fractional delay differential equations.

ϕ: History function

Construct a fractional delayed differential equation problem.

FractionalDiffEq.FDDEMatrixProblem — Type

FDDEMatrixProblem(α, τ, A, B, f, x0, tspan)

Construct a fractional matrix differential equation with delay with general form:

\[D_{t_0}^\alpha\textbf{x}(t)=\textbf{A}(t)\textbf{x}(t)+\textbf{B}(t)\textbf{x}(t-\tau)+\textbf{f}(t)\]

FractionalDiffEq.FDDESystem — Type

FDDESystem(f, ϕ, α, τ, T)

Construct system of fractional delay differential equations problem.

DODE Problem

FractionalDiffEq.DODEProblem — Type

DODEProblem(parameters, orders, interval, tspan, rightfun)

Define distributed order differential equation problem.

Fractional Difference Problem

Missing docstring.

Missing docstring for FractionalDiffEq.FractionalDifferenceProblem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FractionalDiffEq.FractionalDifferenceSystem. Check Documenter's build log for details.

FIE Problem

Missing docstring.

Missing docstring for FractionalDiffEq.FIEProblem. Check Documenter's build log for details.