Problem Types

Problem Types

FODE Problem

FPDE Problem

FDDEProblem

DODEProblem

FODESystem

Fractional Difference Problem

FIEProblem

The general fractional differential equations problem type:

FractionalDiffEq.FDEProblem — Type

FDEProblem

General type for all kinds of problems in FractionalDiffEq.jl.

FODE Problem

Fractional ordinary problems type, we can classified them as Single-Term and Multi-Term problems:

\[D^{\alpha}y(t)=f(t, y)\]

\[\sum_{s=0}^pc_sD^{\beta_s}y=f\]

FractionalDiffEq.SingleTermFODEProblem — Type

SingleTermFODEProblem(f, α, u0, T)

Define a single term fractional ordinary differential equation, there are only one term in this problem.

FractionalDiffEq.MultiTermsFODEProblem — Type

MultiTermsFODEProblem(parameters, orders, rightfun, u0, T)

MultiTermsFODEProblem(parameters, orders, rightfun, rparameters, rorders, u0, T)

Define a multi-terms fractional ordinary differential equation.

FPDE Problem

FractionalDiffEq.FPDEProblem — Type

FPDEProblem(α, β, T, M, N)

Define fractional order partial differential equations problem

FDDEProblem

FractionalDiffEq.FDDEProblem — Type

FDDEProblem(f, ϕ, α, τ, T)
FDDEProblem(f, ϕ, α, τ, t0, T)

Construct a fractional delayed differential equation problem.

FractionalDiffEq.FDDESystem — Type

FDDESystem(f, ϕ, α, τ, T)

Construct system of fractional delay differential equations problem.

DODEProblem

FractionalDiffEq.DODEProblem — Type

DODEProblem(parameters, orders, interval, tspan, rightfun)

Define a single term distributed order differential equation problem.

FODESystem

FractionalDiffEq.FODESystem — Type

FODESystem(f, α, x0)

Define system of fractional differential equations problem.

Fractional Difference Problem

FractionalDiffEq.FractionalDifferenceProblem — Type

FractionalDifferenceProblem(f, α, x0)

Define fractional difference equation problems.

FIEProblem

FractionalDiffEq.FIEProblem — Type

FIEProblem(parameters, orders, rightfun, tspan)

Define fractional integral equation problems.