Problem Types

Problem Types

FODE Problem

FPDE Problem

FDDEProblem

DODEProblem

FODESystem

The general fractional differential equations problem type:

Missing docstring.

Missing docstring for FractionalDiffEq.FDEProblem. Check Documenter's build log for details.

FODE Problem

Fractional ordinary problems type, we can classified them as Single-Term and Multi-Term problems:

\[D^{\alpha}y(t)=f(t, y)\]

\[\sum_{s=0}^pc_sD^{\beta_s}y=f\]

FractionalDiffEq.SingleTermFODEProblem — Type

SingleTermFODEProblem(f, α, h)

Define a single term fractional ordinary differential equation, there are only one term in this problem.

FractionalDiffEq.MultiTermsFODEProblem — Type

MultiTermsFODEProblem(parameters, orders, rightfun)

MultiTermsFODEProblem(parameters, orders, rparameters, rorders)

Define a multi-terms fractional ordinary differential equation.

FPDE Problem

FractionalDiffEq.FPDEProblem — Type

FPDEProblem(α, β, T, M, N)

Define fractional order partial differential equations problem

FDDEProblem

FractionalDiffEq.FDDEProblem — Type

FDDEProblem(f, ϕ, α, τ)

Construct a fractional delayed differential equation peoblem.

DODEProblem

FractionalDiffEq.DODEProblem — Type

SingleTermDODEProblem(ω, t, h, B, rightfun)

Define a single term distributed order differential equation problem.

FODESystem

FractionalDiffEq.FODESystem — Type

FODESystem(f, α, x0)

Define system of fractional differential equations