Problem Types

Problem Types

FODE Problem

FPDE Problem

FDDE Problem

DODE Problem

Fractional Difference Problem

FIE Problem

The general fractional differential equations problem type:

FractionalDiffEq.FDEProblem — Type

FDEProblem

General type for all kinds of problems in FractionalDiffEq.jl.

FODE Problem

Fractional ordinary problems type, we can classified them as Single-Term and Multi-Term problems:

\[D^{\alpha}y(t)=f(t, y)\]

\[\sum_{s=0}^pc_sD^{\beta_s}y=f\]

FractionalDiffEq.SingleTermFODEProblem — Type

SingleTermFODEProblem(f, α, u0, tspan)

Define a single term fractional ordinary differential equation, there are only one fractional differential operator in this problem.

FractionalDiffEq.MultiTermsFODEProblem — Type

MultiTermsFODEProblem(parameters, orders, rightfun, u0, T)

MultiTermsFODEProblem(parameters, orders, rightfun, rparameters, rorders, u0, T)

Define a multi-terms fractional ordinary differential equation.

FractionalDiffEq.FODESystem — Type

FODESystem(f, α, u0, tspan, p)

Define system of fractional differential equations.

FFODE Problem

Missing docstring.

Missing docstring for FractionalDiffEq.FFODEProblem. Check Documenter's build log for details.

FPDE Problem

FractionalDiffEq.FPDEProblem — Type

FPDEProblem(α, β, T, M, N)

Define fractional order partial differential equations problem

FDDE Problem

FractionalDiffEq.FDDEProblem — Type

FDDEProblem(f, ϕ, α, τ, tspan)

Construct a fractional delayed differential equation problem.

FractionalDiffEq.FDDEMatrixProblem — Type

FDDEMatrixProblem(α, τ, A, B, f, x0, tspan)

Construct a fractional matrix differential equation with delay with general form:

\[D_{t_0}^\alpha\textbf{x}(t)=\textbf{A}(t)\textbf{x}(t)+\textbf{B}(t)\textbf{x}(t-\tau)+\textbf{f}(t)\]

FractionalDiffEq.FDDESystem — Type

FDDESystem(f, ϕ, α, τ, T)

Construct system of fractional delay differential equations problem.

DODE Problem

FractionalDiffEq.DODEProblem — Type

DODEProblem(parameters, orders, interval, tspan, rightfun)

Define distributed order differential equation problem.

Fractional Difference Problem

FractionalDiffEq.FractionalDifferenceProblem — Type

FractionalDifferenceProblem(f, α, x0)

Define fractional difference equation problems.

FractionalDiffEq.FractionalDifferenceSystem — Type

FractionalDifferenceSystem(f, α, u0)

Define Fractional Difference System with the general constructure:

\[{^G\nabla_k^\alpha x(k+1)}=f(x(k))\]

With given initial condition $x(i)$.

FIE Problem

Missing docstring.

Missing docstring for FractionalDiffEq.FIEProblem. Check Documenter's build log for details.